name_request_logic_adapter refactored
