+++ src/chatbot/popular_questions_chatbot/choice_algorithm.py
